pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-executor>=1.0.0",
    "keras>=3.13.2",
    "pillow>=12.1.1",
    "torch>=2.10.0",
    "torch-xla>=2.9.0",
]
